Moving vtgate's gateway to its own package.
This includes the aggregator code and the discovery gateway
implementation.
